+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Ignore [jsx `__source`](https://babeljs.io/docs/plugins/transform-react-jsx-source/) attribute.
+
 ## [1.1.0] - 2018-02-07
 
 * Compatibility with Hyperapp [v1.1.0](https://github.com/hyperapp/hyperapp/releases/tag/1.1.0)